requirements.txt
bokeh==2.4.3
Cython==0.29.30
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
numpy==1.23.0
packaging==21.3
pandas==1.4.3
Pillow==9.1.1
protobuf==4.21.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
six==1.16.0
tornado==6.1
typing_extensions==4.2.0
